# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"PaymentRecordDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"sessionId": {
						"type": "number",
						"format": "double"
					},
					"consoleId": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "string"
					},
					"paymentMethodId": {
						"type": "string"
					},
					"paymentMethodName": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"isCash": {
						"type": "boolean"
					},
					"staff": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"paymentMethodId",
					"paymentMethodName",
					"amount",
					"isCash",
					"staff"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PaymentRecordDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentSplitDto": {
				"properties": {
					"paymentMethodId": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"paymentMethodId",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProcessPaymentDto": {
				"properties": {
					"sessionId": {
						"type": "number",
						"format": "double"
					},
					"consoleId": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "string"
					},
					"payments": {
						"items": {
							"$ref": "#/components/schemas/PaymentSplitDto"
						},
						"type": "array"
					},
					"staff": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"payments"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaymentMethodDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isCash": {
						"type": "boolean"
					},
					"isProtected": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"displayOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"nameAr",
					"type",
					"isCash",
					"isProtected",
					"isActive",
					"displayOrder"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentMethodListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PaymentMethodDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentMethodResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/PaymentMethodDto"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePaymentMethodDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isCash": {
						"type": "boolean"
					},
					"displayOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"nameAr"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePaymentMethodDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isCash": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"displayOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GamepadDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"assignedTo": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"number",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GamepadListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GamepadDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GamepadResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/GamepadDto"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ControllerActionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserPublicDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"username",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponseDto": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"token": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserPublicDto"
					}
				},
				"required": [
					"success",
					"token",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequestDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserActionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CurrentUserResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"user": {
						"$ref": "#/components/schemas/UserPublicDto"
					}
				},
				"required": [
					"success",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserPublicDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccountDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playstation-system-backend",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/payments": {
			"get": {
				"operationId": "GetPayments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sessionId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "consoleId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isCash",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "ProcessPayments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProcessPaymentDto"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/payment-methods": {
			"get": {
				"operationId": "GetAllPaymentMethods",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePaymentMethod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentMethodDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/payment-methods/{id}": {
			"get": {
				"operationId": "GetPaymentMethodById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePaymentMethod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePaymentMethodDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePaymentMethod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/controllers": {
			"get": {
				"operationId": "GetAllControllers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GamepadListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Controllers"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GamepadResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Controllers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GamepadDto"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/controllers/{id}": {
			"delete": {
				"operationId": "DeleteController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ControllerActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Controllers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequestDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrentUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/auth/accounts": {
			"get": {
				"operationId": "GetAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/AccountDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}